Implement updating thumbnails automatically

+++ Dev/Cpp/EditorCommon/Platform/FileSystem.h
@@ -11,7 +11,7 @@ class FileSystem
 public:
 	static bool GetIsFile(const std::u16string& path);
 	static bool GetIsDirectory(const std::u16string& path);
-	static int32_t GetLastWriteTime(const std::u16string& path);
+	static uint64_t GetLastWriteTime(const std::u16string& path);
 	static int32_t GetFileSize(const std::u16string& path);
 };
 } // namespace altseed
